json resposne

+++ src/main/java/com/dit/ebay/xml_model/xml_adapters/XMLRatingAdapter.java
@@ -2,13 +2,13 @@ package com.dit.ebay.xml_model.xml_adapters;
 
 import javax.xml.bind.annotation.adapters.XmlAdapter;
 
-public class XMLRatingAdapter extends XmlAdapter<String, Double> {
+public class XMLRatingAdapter extends XmlAdapter<String, Long> {
 
-    public Double unmarshal(String val) throws Exception {
-        return Double.parseDouble(val);
+    public Long unmarshal(String val) throws Exception {
+        return Long.parseLong(val);
     }
 
-    public String marshal(Double val) throws Exception {
+    public String marshal(Long val) throws Exception {
         return val.toString();
     }
 }
